moved GameCore class to screen package
