+++ core/src/miniventure/game/item/Item.java
@@ -1,6 +1,6 @@
 package miniventure.game.item;
 
-import miniventure.game.GameCore;
+import miniventure.game.screen.GameCore;
 import miniventure.game.world.WorldObject;
 import miniventure.game.world.entity.mob.Player;
