+++ core/src/miniventure/game/item/ToolType.java
@@ -1,6 +1,6 @@
 package miniventure.game.item;
 
-import miniventure.game.GameCore;
+import miniventure.game.screen.GameCore;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
