+++ core/src/miniventure/game/world/tile/TileTouchCheck.java
@@ -1,6 +1,6 @@
 package miniventure.game.world.tile;
 
-import miniventure.game.GameCore;
+import miniventure.game.screen.GameCore;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Pixmap;
